Update of Templates (#27)
* Added regulatory references for product certification

* UEP Clarification in Process Steps

UEP creation was mentioned in two process steps -> merged to one (step 3). Same for SW safety classification.

* Removed software requirements review from step 4

Review of SWR in step 5 instead of step 4 (redundant)

* minor wording update

* Updated PRRC / MDSO and referenced SOP vigilance

* Avoiding negative question in supplier list

* Added AFAP approach to RMP

AFAP approach based on 24971:2020
Priority of risk controls based on 14971:2019

Both were only part of risk management report (RMR) so far, but not part of RMP template.

* Quality objectives can be part of management review

Defined two options for the implementation of quality objectives: as part of short-term goals or as part of the annual management review (KPIs).

* Flag outsourced processes

Added pointer to para. 4

* Removed short-term goals

Added to implementation options for quality objectives as part of the quality manual process. As a result, removed the concept of short-term goals here.

* Correcting typo

* formatting

---------

Co-authored-by: Oliver Eidel <oli@eidel.net>